add Progressar component
