+++ src/components/MapDrawer.js
@@ -13,6 +13,7 @@ import log from 'loglevel';
 import WalletInfo from './WalletInfo';
 import CustomizedTabs from './common/CustomizedTabs';
 import TokenCard from './TokenCard';
+import LinearProgressBar from './common/LinearProgressBar';
 
 const style = (theme) => ({
   map: {
@@ -115,6 +116,7 @@ function MapDrawer(props) {
             <WalletInfo />
             <CustomizedTabs tab1="Tokens" tab2="Impact" />
             <TokenCard />
+            <LinearProgressBar value={70} width="90%" height="24px" />
           </Grid>
         </Paper>
       </SwipeableDrawer>
